+++ package.json
@@ -28,9 +28,11 @@
     "prettier": "^2.2.1",
     "serverless-offline": "^6.9.0",
     "webpack": "^5.28.0",
-    "webpack-cli": "^4.6.0"
+    "webpack-cli": "^4.6.0",
+    "webpack-node-externals": "^2.5.2"
   },
   "dependencies": {
+    "dotenv": "^8.2.0",
     "pg": "^8.5.1"
   }
 }
