+++ packages/product/webpack.config.js
@@ -1,6 +1,5 @@
 const path = require('path');
 const glob = require('glob');
-
 const baseConfig = require('../../webpack-base.config');
 
 const entries = glob.sync('./handlers/*.js').reduce(
@@ -13,7 +12,10 @@ const entries = glob.sync('./handlers/*.js').reduce(
 
 module.exports = {
   ...baseConfig,
-  entry: entries,
+  entry: {
+    ...entries,
+    'populate-products.js': './db/postgres/scripts/populate-products.js',
+  },
   output: {
     ...baseConfig.output,
     path: path.join(__dirname, 'build'),
